add activity_film
