+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.edurda77.filmlibrary" >
+    package="com.edurda77.filmlibrary">
 
     <application
         android:allowBackup="true"
@@ -8,7 +8,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.FilmLibrary" >
+        android:theme="@style/Theme.FilmLibrary">
+        <activity
+            android:name=".domain.FilmActivity"
+            android:exported="false" />
         <activity
             android:name=".ui.AboutActivity"
             android:exported="false" />
@@ -20,7 +23,7 @@
             android:exported="false" />
         <activity
             android:name=".ui.MainActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
