+++ app/src/main/java/com/edurda77/filmlibrary/domain/FilmActivity.kt
@@ -1,0 +1,12 @@
+package com.edurda77.filmlibrary.domain
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.edurda77.filmlibrary.R
+
+class FilmActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_film)
+    }
+}
