+++ app/src/main/java/com/edurda77/filmlibrary/ui/MainActivity.kt
@@ -11,10 +11,12 @@ import androidx.recyclerview.widget.RecyclerView
 import com.edurda77.filmlibrary.R
 import com.edurda77.filmlibrary.domain.FilmGenre
 import com.edurda77.filmlibrary.domain.Movie
+import android.widget.Toast
+import com.edurda77.filmlibrary.ui.MovieAdapter.OnStateClickListener
 
 
 private var toolbar: Toolbar? = null
-
+val movies: ArrayList<Movie> = ArrayList()
 class MainActivity : AppCompatActivity() {
     private var action = listOf(
         Movie("Терминатор", "action", "120 min",10.0, 1984, 1.0, 3.3, "fgfgfgfgfg"),
@@ -36,10 +38,12 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         //var movie : Movie
+
         super.onCreate(savedInstanceState)
 
         setContentView(R.layout.activity_main)
         setToolbar()
+
         setOotRecycledView()
 
     }
@@ -50,8 +54,10 @@ class MainActivity : AppCompatActivity() {
     }
     fun setOotRecycledView() {
         val recyclerView: RecyclerView = findViewById(R.id.out_recycled_view)
+
         recyclerView.layoutManager = LinearLayoutManager(this,LinearLayoutManager.VERTICAL, false)
-        recyclerView.adapter = SecondAdapter(ganre)
+        recyclerView.adapter = OutAdapter(ganre)
+
     }
