+++ app/src/main/res/layout/activity_film.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/purple_500"
+    android:orientation="vertical"
+    android:layout_margin="@dimen/margin_half">
+    <TextView
+        style="@style/title_movie_view"
+        android:id="@+id/title_movie"
+        android:text="@string/ttile_movie"
+        />
+    <ImageView
+        android:id="@+id/picture_movie"
+        android:layout_width="@dimen/picture_size"
+        android:layout_height="@dimen/picture_size"
+        android:layout_gravity="center"/>
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/year_movie"
+        android:text="@string/year_movie"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/rang_movie"
+        android:text="@string/rang_movie"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/ganre_movie"
+        android:text="@string/ganre"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/duration_movie"
+        android:text="@string/duration"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/budget_movie"
+        android:text="@string/budget"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/revenue_movie"
+        android:text="@string/revenue"
+        />
+    <TextView
+        style="@style/year_movie_view"
+        android:id="@+id/summary_movie"
+        android:text="@string/summary"
+        />
+
+</LinearLayout>
