+++ app/src/main/res/values-night/strings.xml
@@ -7,4 +7,8 @@
     <string name="menu_custom">Настройки</string>
     <string name="menu_about">О приложении</string>
     <string name="rang_movie">Рейтинг фильма</string>
+    <string name="revenue">Кассовые сборы</string>
+    <string name="budget">Бюджет</string>
+    <string name="duration">Продолжительность</string>
+    <string name="ganre">Жанр</string>
 </resources>
