+++ app/src/main/res/values/styles.xml
@@ -15,7 +15,7 @@
     <style name="title_movie_view">
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
-        <item name="android:textColor">@color/green</item>
+        <item name="android:textColor">@color/black</item>
         <item name="android:layout_margin">@dimen/margin_half</item>
         <item name="android:textSize">@dimen/margin_twin</item>
         <item name="android:gravity">center</item>
